+++ src/module/setupModules.ts
@@ -3,7 +3,7 @@ import { log } from "../midi-qol.js";
 import { TroubleShooter } from "./apps/TroubleShooter.js";
 import { configSettings } from "./settings.js";
 
-export const DAE_REQUIRED_VERSION = "12.0.7";
+export const DAE_REQUIRED_VERSION = "12.0.8";
 export const REQUIRED_MODULE_VERSIONS = {
   "about-time": "0.0", 
   "anonymous": "0.0.0",
@@ -23,7 +23,7 @@ export const REQUIRED_MODULE_VERSIONS = {
   "monks-tokenbar": "1.0.55",
   "multilevel-tokens": "1.6.0",
   "perceptive": "3.2.1",
-  // "ready-set-roll-5e": "1.2.0",
+  "ready-set-roll-5e": "3.4.3",
   "simbuls-cover-calculator":  "1.0.2",
   "socketlib": "0.0",
   "times-up": "11.3.0",